# backend_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mine Worker Safety Management System API",
    "description": "Backend API for Role-Based Access Control, Shift Monitoring, Hazard Reporting, Live Tracking, and SOS Alerts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/shifts/start": {
      "post": {
        "tags": [
          "Shifts"
        ],
        "summary": "Start Shift",
        "operationId": "start_shift_api_shifts_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/shifts/end": {
      "post": {
        "tags": [
          "Shifts"
        ],
        "summary": "End Shift",
        "operationId": "end_shift_api_shifts_end_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/shifts/active": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "Get Active Shift",
        "operationId": "get_active_shift_api_shifts_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ShiftOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Active Shift Api Shifts Active Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/shifts/history": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "Get Shift History",
        "operationId": "get_shift_history_api_shifts_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response Get Shift History Api Shifts History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/shifts/admin/history": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "Get All Shift History",
        "operationId": "get_all_shift_history_api_shifts_admin_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response Get All Shift History Api Shifts Admin History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/sos/trigger": {
      "post": {
        "tags": [
          "SOS Emergency"
        ],
        "summary": "Trigger Sos",
        "operationId": "trigger_sos_api_sos_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SOSAlertCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SOSAlertOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sos/active": {
      "get": {
        "tags": [
          "SOS Emergency"
        ],
        "summary": "Get Active Sos",
        "operationId": "get_active_sos_api_sos_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SOSAlertOut"
                  },
                  "type": "array",
                  "title": "Response Get Active Sos Api Sos Active Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sos/history": {
      "get": {
        "tags": [
          "SOS Emergency"
        ],
        "summary": "Get Sos History",
        "operationId": "get_sos_history_api_sos_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SOSAlertOut"
                  },
                  "type": "array",
                  "title": "Response Get Sos History Api Sos History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sos/{sos_id}/status": {
      "put": {
        "tags": [
          "SOS Emergency"
        ],
        "summary": "Update Sos Status",
        "operationId": "update_sos_status_api_sos__sos_id__status_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sos_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sos Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SOSAlertUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SOSAlertOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SOSAlertCreate": {
        "properties": {
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Longitude"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type",
            "default": "SOS_TRIGGERED"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "SOSAlertCreate"
      },
      "SOSAlertOut": {
        "properties": {
          "latitude": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Latitude"
          },
          "longitude": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Longitude"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type",
            "default": "SOS_TRIGGERED"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "worker_id": {
            "type": "integer",
            "title": "Worker Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "id",
          "worker_id",
          "status",
          "timestamp"
        ],
        "title": "SOSAlertOut"
      },
      "SOSAlertUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SOSAlertUpdate"
      },
...
      "ShiftOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "worker_id": {
            "type": "integer",
            "title": "Worker Id"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "total_hours": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Hours"
          },
          "attendance_status": {
            "type": "string",
            "title": "Attendance Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "worker_id",
          "start_time",
          "attendance_status",
          "created_at"
        ],
        "title": "ShiftOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}